Adding a Randramb mode and Nicknames mode.
To activate a a mode - you need to add a key to your local stroage using a browsers dev tools.
For Randramb mode add a key randrambMode with a value of 1.
For Nicknames mode add a key nicknamesMode with a value of 1.

Please note that I do not support the mess that may result from either mode.